// src/app/manager-smart-form/card-services/shared/models/card-service-common.ts
export interface BranchInfo {
  id: string;
  parentId: string;
  name: string;
  code: string;
  phone: string;
  fax: string;
  contactName: string;
  email: string;
  address: string;
  cityCode: string;
  districtCode: string;
  displayPriority: string;
  statusCode: string;
  createdDate: string;
  createdBy: string;
  modifiedDate: string;
  modifiedBy: string;
  branchCodeName: string;
}